requirements.txt
Flask>=3.0.0
Pillow>=10.0.0
opencv-python-headless>=4.10.0.0
numpy>=1.26.0
